Just a warning
git-svn-id: https://svn.code.sf.net/p/bzflag/code/trunk/bzflag@4370 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ src/bzfs/bzfs.cxx
@@ -2273,7 +2273,7 @@ static void addPlayer(int playerIndex)
 
       // see if the player's choice was a weak team
       bool foundTeam = false;
-      for (int i = 0; i < minIndex.size(); i++) {
+      for (int i = 0; i < (int) minIndex.size(); i++) {
 	if (minIndex[i] == (TeamColor)t) {
 	  foundTeam = true;
 	  break;
